Add slumber_util subcrate
Moving some stuff into slumber_util breaks the config->core dependency, which reduces the amount of code that needs to recompile when making changes to core.
