+++ crates/core/src/db/tests.rs
@@ -1,8 +1,9 @@
 use super::*;
-use crate::{test_util::Factory, util::paths::get_repo_root};
+use crate::test_util::Factory;
 use indexmap::IndexMap;
 use itertools::Itertools;
 use rstest::{fixture, rstest};
+use slumber_util::paths::get_repo_root;
 use std::collections::HashMap;
 
 impl CollectionDatabase {
