+++ crates/core/src/http.rs
@@ -49,7 +49,6 @@ use crate::{
     db::CollectionDatabase,
     http::{content_type::ContentType, curl::CurlBuilder},
     template::{Template, TemplateContext},
-    util::ResultTraced,
 };
 use anyhow::Context;
 use bytes::Bytes;
@@ -65,9 +64,10 @@ use reqwest::{
     header::{self, HeaderMap, HeaderName, HeaderValue},
     multipart::{Form, Part},
 };
-use serde::{Deserialize, Serialize};
-use std::collections::HashSet;
-use tracing::{info, info_span};
+use slumber_config::HttpEngineConfig;
+use slumber_util::ResultTraced;
+use std::{collections::HashSet, error::Error};
+use tracing::{error, info, info_span};
 
 const USER_AGENT: &str = concat!("slumber/", env!("CARGO_PKG_VERSION"));
 
@@ -361,38 +361,6 @@ impl Default for HttpEngine {
     }
 }
 
-#[derive(Debug, Serialize, Deserialize)]
-#[serde(default)]
-pub struct HttpEngineConfig {
-    /// TLS cert errors on these hostnames are ignored. Be careful!
-    pub ignore_certificate_hosts: Vec<String>,
-    /// Request/response bodies over this size are treated differently, for
-    /// performance reasons
-    pub large_body_size: usize,
-    /// Enable/disable persistence for _all_ requests? The CLI should override
-    /// this based on the absence/presence of the `--persist` flag
-    pub persist: bool,
-}
-
-impl HttpEngineConfig {
-    /// Is the given size (e.g. request or response body size) larger than the
-    /// configured "large" body size? Large bodies are treated differently, for
-    /// performance reasons.
-    pub fn is_large(&self, size: usize) -> bool {
-        size > self.large_body_size
-    }
-}
-
-impl Default for HttpEngineConfig {
-    fn default() -> Self {
-        Self {
-            ignore_certificate_hosts: Default::default(),
-            large_body_size: 1000 * 1000, // 1MB
-            persist: true,
-        }
-    }
-}
-
 impl RequestSeed {
     /// Run the given future and convert any error into [RequestBuildError]
     async fn run_future<T>(
@@ -468,7 +436,7 @@ impl RequestTicket {
                 end_time,
                 error: error.into(),
             })
-            .traced(),
+            .inspect_err(|err| error!(error = err as &dyn Error)),
         }
     }
 }
