+++ crates/core/src/test_util.rs
@@ -2,15 +2,16 @@
 
 use crate::{
     collection::{ChainSource, HasId},
-    http::{HttpEngine, HttpEngineConfig},
+    http::HttpEngine,
     template::{Prompt, Prompter, Select},
-    util::{ResultTraced, paths::get_repo_root},
 };
 use anyhow::Context;
 use derive_more::Deref;
 use indexmap::IndexMap;
 use reqwest::header::{HeaderMap, HeaderName, HeaderValue};
 use rstest::fixture;
+use slumber_config::HttpEngineConfig;
+use slumber_util::{ResultTraced, paths::get_repo_root};
 use std::{
     env, fs,
     path::PathBuf,
@@ -180,51 +181,4 @@ pub fn header_map<'a>(
             )
         })
         .collect()
-}
-
-/// Assert a result is the `Err` variant, and the stringified error contains
-/// the given message
-#[macro_export]
-macro_rules! assert_err {
-    ($e:expr, $msg:expr) => {{
-        use itertools::Itertools as _;
-
-        let msg = $msg;
-        // Include all source errors so wrappers don't hide the important stuff
-        let error: anyhow::Error = $e.unwrap_err().into();
-        let actual = error.chain().map(ToString::to_string).join(": ");
-        assert!(
-            actual.contains(msg),
-            "Expected error message to contain {msg:?}, but was: {actual:?}"
-        )
-    }};
-}
-
-/// Assert the given expression matches a pattern and optional condition.
-/// Additionally, evaluate an expression using the bound pattern. This can be
-/// used to apply additional assertions inline, or extract bound values to use
-/// in subsequent statements.
-#[macro_export]
-macro_rules! assert_matches {
-    ($expr:expr, $pattern:pat $(if $condition:expr)? $(,)?) => {
-        $crate::assert_matches!($expr, $pattern $(if $condition)? => ());
-    };
-    ($expr:expr, $pattern:pat $(if $condition:expr)? => $output:expr $(,)?) => {
-        match $expr {
-            // If a conditional was given, check it. This has to be a separate
-            // arm to prevent borrow fighting over the matched value
-            $(value @ $pattern if !$condition => {
-                panic!(
-                    "Value {value:?} does not match condition {condition}",
-                    condition = stringify!($condition),
-                );
-            })?
-            #[allow(unused_variables)]
-            $pattern => $output,
-            value => panic!(
-                "Unexpected value {value:?} does not match pattern {expected}",
-                expected = stringify!($pattern),
-            ),
-        }
-    };
 }
