+++ crates/import/Cargo.toml
@@ -22,6 +22,7 @@ serde = {workspace = true}
 serde_json = {workspace = true}
 serde_yaml = {workspace = true}
 slumber_core = {workspace = true}
+slumber_util = {workspace = true}
 strum = {workspace = true}
 thiserror = {workspace = true}
 tracing = {workspace = true}
