+++ crates/tui/Cargo.toml
@@ -31,6 +31,7 @@ serde_yaml = {workspace = true}
 shell-words = "1.1.0"
 slumber_config = {workspace = true}
 slumber_core = {workspace = true}
+slumber_util = {workspace = true}
 strum = {workspace = true}
 tokio = {workspace = true, features = ["macros", "signal", "tracing"]}
 tokio-util = "0.7.13"
