+++ crates/util/src/lib.rs
@@ -1,0 +1,80 @@
+//! Common utilities that aren't specific to one other subcrate and are unlikely
+//! to change frequently. The main purpose of this is to pull logic out of the
+//! core crate, because that one changes a lot and requires constant
+//! recompilation.
+
+pub mod paths;
+#[cfg(feature = "test")]
+pub mod test_util;
+
+use serde::de::DeserializeOwned;
+use std::{fmt::Debug, io::Read, ops::Deref};
+use tracing::error;
+
+/// A static mapping between values (of type `T`) and labels (strings). Used to
+/// both stringify from and parse to `T`.
+pub struct Mapping<'a, T: Copy>(&'a [(T, &'a [&'a str])]);
+
+impl<'a, T: Copy> Mapping<'a, T> {
+    /// Construct a new mapping
+    pub const fn new(mapping: &'a [(T, &'a [&'a str])]) -> Self {
+        Self(mapping)
+    }
+
+    /// Get a value by one of its labels
+    pub fn get(&self, s: &str) -> Option<T> {
+        for (value, strs) in self.0 {
+            for other_string in *strs {
+                if *other_string == s {
+                    return Some(*value);
+                }
+            }
+        }
+        None
+    }
+
+    /// Get the label mapped to a value. If it has multiple labels, use the
+    /// first. Panic if the value has no mapped labels
+    pub fn get_label(&self, value: T) -> &str
+    where
+        T: Debug + PartialEq,
+    {
+        let (_, strings) = self
+            .0
+            .iter()
+            .find(|(v, _)| v == &value)
+            .unwrap_or_else(|| panic!("Unknown value {value:?}"));
+        strings
+            .first()
+            .unwrap_or_else(|| panic!("No mapped strings for value {value:?}"))
+    }
+
+    /// Get all available mapped strings
+    pub fn all_strings(&self) -> impl Iterator<Item = &str> {
+        self.0
+            .iter()
+            .flat_map(|(_, strings)| strings.iter().copied())
+    }
+}
+/// Extension trait for [Result]
+pub trait ResultTraced<T, E>: Sized {
+    /// If this is an error, trace it. Return the same result.
+    fn traced(self) -> Self;
+}
+
+impl<T> ResultTraced<T, anyhow::Error> for anyhow::Result<T> {
+    fn traced(self) -> Self {
+        self.inspect_err(|err| error!(error = err.deref()))
+    }
+}
+
+/// Parse bytes from a reader into YAML. This will merge any anchors/aliases.
+pub fn parse_yaml<T: DeserializeOwned>(reader: impl Read) -> anyhow::Result<T> {
+    // Two-step parsing is required for anchor/alias merging
+    let deserializer = serde_yaml::Deserializer::from_reader(reader);
+    let mut yaml_value: serde_yaml::Value =
+        serde_path_to_error::deserialize(deserializer)?;
+    yaml_value.apply_merge()?;
+    let output = serde_path_to_error::deserialize(yaml_value)?;
+    Ok(output)
+}
